requirements.txt
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.22.0
autopep8==1.3.4
beautifulsoup4==4.5.3
bleach==3.1.4
cachetools==3.1.1
cairocffi==1.0.2
CairoSVG==2.4.2
certifi==2018.4.16
cffi==1.10.0
chardet==3.0.2
cryptography==1.8.1
cssselect==1.0.1
cssselect2==0.2.1
decorator==4.0.11
Django==1.11.28
django-anymail==7.1.0
django-autocomplete-light==3.4.1
django-common-helpers==0.9.0
django-compressor-postcss==0.8.2
django-contrib-comments==1.8.0
django-debug-toolbar==1.9.1
django-multiurl==1.1.0
django-nested-admin==3.2.4
django_compressor==2.3
djangorestframework==3.7.0
django-webpack-loader==0.7.0
emoji==0.5.4
Fabric3==1.13.1.post1
filebrowser-safe==0.5.0
future==0.16.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
grappelli-safe==0.5.0
html5lib==0.999999999
httplib2==0.13.1
idna==2.5
ipdb==0.10.3
ipython==5.3.0
ipython-genutils==0.2.0
lxml==3.7.3
Mezzanine==4.3.1
mysqlclient==1.3.10
oauthlib==2.0.2
olefile==0.44
packaging==16.8
paramiko==2.1.6
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.2.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.3.1
pycparser==2.17
Pygments==2.2.0
pylabels==1.2.1
pyparsing==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.4
python-crontab==2.1.1
python-dateutil==2.6.0
pytz==2016.10
reportlab==3.4.0
requests==2.20.0
requests-oauthlib==0.8.0
rollbar==0.14.7
rsa==4.0
signrequest-python-client==2.1.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
stripe==1.82.2
tinycss==0.4
tinycss2==1.0.2
traitlets==4.3.2
tzlocal==1.3
uritemplate==3.0.0
urllib3==1.24.2
wcwidth==0.1.7
WeasyPrint==51
webencodings==0.5
xhtml2pdf==0.0.6
celery==5.2.3
redis==4.1.2
